--- src/App.css
@@ -1,18 +1,0 @@
-.App {
-  text-align: center;
-  background-color: #44C572;
-  height: 100%
-}
-
-.Body {
-  background-color: #33AE5E;
-  pointer-events: none;
-  font-weight: bolder;
-  padding: 50px 50px;
-}
-
-.inner {
-  background-color: #6DC18A;
-  height: 200px;
-  margin-bottom: 50px;
-}
